Rename `pytorch` to `src` as the future main source folder
